CloudComputing/week6/miniproject2/Clasification_and_Guestbook_elements/Clasificador/requirements.txt
flask
gluoncv
matplotlib
mxnet
requests
Pillow
